refactor: split Layout and AiService god files into focused modules
- frontend: Layout.tsx becomes 18-line router switch; public header/nav
  extracted to PublicLayout.tsx, admin sidebar/nav to AdminLayout.tsx,
  BrandFallbackTile extracted to its own file
- backend: AiService becomes 60-line orchestrator; description generation
  to AiDescriptionService, Facebook post to AiFacebookService, image
  analysis to AiImageService (all under ai/services/)
- Public API of both Layout and AiService unchanged; no behavior changes
